refactor: 불필요한 외부 라이브러리 제거 및 stream으로 대체
