+++ src/main/java/controller/RequestController.java
@@ -1,20 +1,21 @@
-package controller;
+package webserver;
 
+import com.github.jknack.handlebars.internal.lang3.ArrayUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import httpMock.CustomHttpRequest;
-import httpMock.CustomHttpResponse;
-import httpMock.constants.ContentType;
-import httpMock.constants.StatusCode;
+import webserver.httpMock.CustomHttpRequest;
+import webserver.httpMock.CustomHttpResponse;
+import webserver.httpMock.constants.ContentType;
+import webserver.httpMock.constants.StatusCode;
 
-public interface RequestController {
-    Logger logger = LoggerFactory.getLogger(RequestController.class);
+public interface RequestService {
+    Logger logger = LoggerFactory.getLogger(RequestHandler.class);
 
     void handleRequest(CustomHttpRequest req, CustomHttpResponse res);
 
     static void NOT_FOUND(CustomHttpResponse res) {
         res.setStatusCode(StatusCode.NOT_FOUND);
         res.setContentType(ContentType.TEXT_PLAIN);
-        res.setBody("404 not found".getBytes());
+        res.addToBody(ArrayUtils.toObject("404 not found".getBytes()));
     }
 }
